requirements.txt
cachetools==4.2.0
certifi==2020.12.5
chardet==4.0.0
google-api-core==1.24.1
google-auth==1.24.0
google-cloud-vision==2.0.0
googleapis-common-protos==1.52.0
grpcio==1.34.0
idna==2.10
libcst==0.3.16
mypy-extensions==0.4.3
numpy==1.19.5
pandas==1.2.0
proto-plus==1.13.0
protobuf==3.14.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
python-dateutil==2.8.1
pytz==2020.5
PyYAML==5.3.1
requests==2.25.1
rsa==4.7
six==1.15.0
typing-extensions==3.7.4.3
typing-inspect==0.6.0
urllib3==1.26.2
